allowing better check on conversation if it is busy or not - main issue is we dont control cause of our spi where a boolean is set (servlet layer? jsf layer?...) so we need to support multiple calls by thread + avoid to serialize the conversation itself but use conversation manager for it
git-svn-id: https://svn.apache.org/repos/asf/openwebbeans/trunk@1648291 13f79535-47bb-0310-9956-ffa450edef68

+++ webbeans-web/src/main/java/org/apache/webbeans/web/context/WebContextsService.java
@@ -173,7 +173,10 @@ public class WebContextsService extends AbstractContextsService
         
         //Destroy singleton context
         endContext(Singleton.class, destroyObject);
-        sharedSingletonContext.destroy();
+        if (sharedSingletonContext != null)
+        {
+            sharedSingletonContext.destroy();
+        }
 
         //Clear saved contexts related with 
         //this servlet context
@@ -438,7 +441,7 @@ public class WebContextsService extends AbstractContextsService
             ConversationImpl owbConversation = (ConversationImpl)conversation;
             owbConversation.updateTimeOut();
             //Other threads can now access propogated conversation.
-            owbConversation.setInUsed(false);
+            owbConversation.iDontUseItAnymore();
         }
     }
